requirements.txt
streamlit
pandas
matplotlib
seaborn
plotly.express
scikit-learn
numpy
joblib
statsmodels
streamlit_option_menu
xgboost